Creates a taglib called BootstrapValidation.
It permits a developer to show a feedback in Bootstrap like-style. The
files aren't full comment yet, because I don't know if there is any
interest in this feature or if it follow the concept style of OmniFaces. 
This is not a component, just a taglib. It will add classes to the
component and its parent (according to its configuration) allowing a css
feedback like Bootstrap. Its totally compatible with <o:skipValidators>
allowing re-validate the appearance of a UIInput without invalidate it.